// Frontend/src/app/models/employee.ts
/* eslint-disable @typescript-eslint/no-explicit-any */
export class Employee {
  _id?: string;
  employeeName: string;
  email: string;
  address: string;
  provinceId: string;
  birthDate: Date;
  password: string;
  photo: string;
  role: string;
  phone: string;
  Province?: any[];
  isworking: boolean;
}